Refactor source tree
