+++ CMakeLists.txt
@@ -2,18 +2,27 @@ cmake_minimum_required(VERSION 3.0)
 project(test)
 
 find_package(xxHash 0.7 CONFIG REQUIRED)
+
 #add_executable(main main.cpp)
-add_executable(l0sampling l0sampling.cpp sketch.cpp update.cpp)
+add_executable(l0sampling
+    l0sampling.cpp
+    l0_sampling/sketch.cpp
+    l0_sampling/update.cpp
+    )
 target_link_libraries(l0sampling PRIVATE xxHash::xxhash)
 # set compiler options
 set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -W -Wall -O3 -std=c++11")
 
 find_package(xxHash 0.7 CONFIG REQUIRED)
+find_package(GTest REQUIRED)
+include_directories(${GTEST_INCLUDE_DIRS})
 
 add_executable(tests
-    tests.cpp
     test/sketch_test.cpp
-    sketch.cpp
-    update.cpp
+    l0_sampling/sketch.cpp
+    l0_sampling/update.cpp
     )
-target_link_libraries(tests PRIVATE xxHash::xxhash criterion)
+target_link_libraries(tests xxHash::xxhash criterion ${GTEST_LIBRARIES} pthread)
+
+include(CTest)
+add_test(testing_suite tests)
