+++ include/stream.h
