+++ l0_sampling/update.cpp
@@ -1,5 +1,5 @@
 #include <iostream>
-#include "update.h"
+#include "../include/update.h"
 
 std::ostream& operator<< (std::ostream &out, const Update &update){
     out << "Index: " << update.index << " Value: " << update.delta;
