--- tests.h
@@ -1,10 +1,0 @@
-#pragma once
-#include "sketch.h"
-
-namespace Tests{
-  //This test
-  void sketch_sum_test(){
-
-  }
-
-}
